Revert unnecessary move
